Refactor threading model (and basically everything)
Now, we only spawn a single thread pool (see the `QTasks._job_handler`
method) in a singleton class accessible by the client request contexts.

Pending jobs are queued in-memory. Current job states and pending
cleanup tasks are also stored in-memory (by default we keep results for
up to 15 minutes). The thread that spawns the pool is responsible for
processing completed jobs, submitting new jobs to the executor and for
cleaning up old results.

Signed-off-by: dorimedini <dorimedini@gmail.com>